requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==8.0.1
Werkzeug==1.0.1

cffi==1.14.3
numpy==1.19.3
pip==20.3.3
wheel==0.36.1
cryptography==3.2.1
colorama==0.4.4
setuptools==47.1.0
torch==1.7.1
six==1.15.0
tensorboard==2.1.1
tensorflow==2.1.0
torchvision==0.4.1
tornado==6.1
matplotlib==3.3.2
scipy==1.4.1
opencv-python==4.4.0.46
mock==4.0.3
PyYAML==5.3.1